+++ 30_sources/SRC_Zupfnoter/src/controller_command_definitions.rb
@@ -294,13 +294,14 @@ C,
             'startpos'     => lambda { {key: "extract.#{@systemstatus[:view]}.startpos", value: $conf['extract.0.startpos']} },
             'subflowlines' => lambda { {key: "extract.#{@systemstatus[:view]}.subflowlines", value: $conf['extract.0.subflowlines']} },
             'produce'      => lambda { {key: "produce", value: $conf['produce']} },
-            'layout'      => lambda { {key:   "extract.#{@systemstatus[:view]}.layout",
+            'layout'       => lambda { {key:   "extract.#{@systemstatus[:view]}.layout",
                                         value: {LINE_THIN:    0.1,
                                                 LINE_MEDIUM:  0.3,
                                                 LINE_THICK:   0.5,
                                                 # all numbers in mm
                                                 ELLIPSE_SIZE: [3.5, 1.7], # radii of the largest Ellipse
                                                 REST_SIZE:    [4, 2]}} }, # radii of the largest Rest Glyph} },
+            'countnotes'           => lambda { {key: "extract.#{@systemstatus[:view]}.countnotes", value: $conf['extract.0.countnotes']} },
             'xx'           => lambda { {key: "xx", value: $conf[]} }
         }